refactor: move data usecases to its own entities folder
