+++ src/data/usecases/account/add-account/db-add-account-usecase-protocols.ts
@@ -2,3 +2,6 @@ export * from '@src/data/protocols/crypto/hasher'
 export * from '@src/data/protocols/db/account/add-account-repository'
 export * from '@src/domain/models/account'
 export * from '@src/domain/usecases/account/add-account'
+export {
+  LoadAccountByEmailRepository,
+} from '@src/data/protocols/db'
