+++ src/data/usecases/account/authentication/db-authentication-usecase-result.ts
