+++ src/main/factories/usecases/add-survey/db-add-survey-usecase-factory.ts
@@ -1,5 +1,5 @@
-import { DbAddSurveyUseCase } from '@src/data/usecases/add-survey/db-add-survey-usecase'
-import { DbAddSurveyResult } from '@src/data/usecases/add-survey/db-add-survey-usecase-result'
+import { DbAddSurveyUseCase } from '@src/data/usecases/survey/add-survey/db-add-survey-usecase'
+import { DbAddSurveyResult } from '@src/data/usecases/survey/add-survey/db-add-survey-usecase-result'
 import { AddSurveyUseCase } from '@src/domain/usecases/survey/add-survey'
 import { SurveyMongoRepository } from '@src/infra/db/mongodb/survey/survey-mongo-repository'
