+++ src/presentation/controllers/account/login/login-controller.test.ts
@@ -5,7 +5,7 @@ import {
   TokenGeneratorError,
   UpdateAccessTokenRepositoryError,
   UserAccessToken
-} from '@src/data/usecases/authentication/db-authentication-usecase-result'
+} from '@src/data/usecases/account/authentication/db-authentication-usecase-result'
 import { InvalidCredentialsError } from '@src/domain/errors'
 import { MissingParamError, ServerError } from '@src/presentation/errors'
 import {
